+++ SLAE/Number_method/SLE_input.h
@@ -6,13 +6,13 @@ class SLE_input
 {
 	std::ifstream file;
 	int n;
-	std::vector < std::vector <float> > a;
-	std::vector < float > b, x;
+	std::vector < std::vector <double> > a;
+	std::vector < double > b, x;
 public:
 	SLE_input(const char* );
-	std::vector < std::vector <float> > getA() { return a; }
-	std::vector < std::vector <float> > getB_X() { return { b, x }; }
-	float getN() { return n; }
+	std::vector < std::vector <double> > getA() { return a; }
+	std::vector < std::vector <double> > getB_X() { return { b, x }; }
+	double getN() { return n; }
 	
 };
